Add Puppetlabs forge modules
Add the code for this modules as well so a complete deployment is
possible with a single checkout.

The rake task pulling and updating these modules must be used to keep
these files up to date. But this allows us to check every change made to
any of these modules before we push them out to important environments.